fix: remediate batch 5 clang-tidy violations (#230)
* fix: remediate batch 5 clang-tidy violations

Signed-bitwise (hicpp-signed-bitwise, ~349 fixes):
- Cast uint8_t/uint16_t to uint32_t before bitwise shifts to
  prevent UB from signed int promotion
- Add U suffix to hex literals used with bitwise operators
- Fix across all modules: sd_message, types.h, e2e_crc,
  tp_segmenter, tp_reassembler, serializer, tcp_transport,
  event_subscriber, net_impl (posix/zephyr/lwip)

Explicit conversions (5 fixes):
- Add explicit to single-argument Pimpl constructors

Qualified auto (3 fixes):
- Use const auto* for pointer declarations in serializer

Error handling (1 fix):
- Replace atoi with strtol in endpoint validation

* fix: address CodeRabbit review findings

- Clamp freshness_timeout_ms to 0xFFFF before narrowing to uint16_t
  to prevent silent truncation for large timeout values
- Widen payload_offset and serialize_tp_header offset parameter from
  uint16_t to uint32_t to support messages larger than 65535 bytes
  (the whole purpose of TP segmentation)

+++ include/platform/lwip/net_impl.h
@@ -44,14 +44,18 @@ static inline int someip_shutdown_socket(someip_socket_t fd) {
 static inline int someip_set_nonblocking(someip_socket_t fd) {
     int flags = lwip_fcntl(fd, F_GETFL, 0);
     if (flags < 0) return -1;
-    return lwip_fcntl(fd, F_SETFL, flags | O_NONBLOCK);
+    return lwip_fcntl(fd, F_SETFL,
+                      static_cast<int>(static_cast<unsigned int>(flags) |
+                                       static_cast<unsigned int>(O_NONBLOCK)));
 }
 
 /** @implements REQ_PLATFORM_LWIP_001 */
 static inline int someip_set_blocking(someip_socket_t fd) {
     int flags = lwip_fcntl(fd, F_GETFL, 0);
     if (flags < 0) return -1;
-    return lwip_fcntl(fd, F_SETFL, flags & ~O_NONBLOCK);
+    return lwip_fcntl(fd, F_SETFL,
+                      static_cast<int>(static_cast<unsigned int>(flags) &
+                                       ~static_cast<unsigned int>(O_NONBLOCK)));
 }
 
 /* ---------- Socket creation & connection ----------------------------------- */
